function-app/requirements.txt
azure-functions
elastic-apm
requests
structlog